added authenticateToken for every routing - now you must send token in every fetch exept login & register
